+++ EDDI/ChangeLog.md
@@ -3,6 +3,8 @@
 Full details of the variables available for each noted event, and VoiceAttack integrations, are available in the individual [event pages](https://github.com/EDCD/EDDI/wiki/Events).
 
 ### 4.0.0-rc1
+  * Inara Responder
+    * Incorporated Inara API updates for Odyssey (except suit loadouts will be implemented later)
   * Speech Responder
     * Events
       * `Backpack` event added